Enemy - Implement Hearing and Sight Feature
Short Explanation
Added CanHideComponent, which must be added to any actor the employees should look/listen for. The component includes a isHidden, which, if true, makes it so the actor cannot be seen no matter how close to the employee. If an actor is seen, the employee will chase them down, and if they lose track, they will return to their original position.

Objects Affected
BP_EnemyEmployee
BP_ThirdPersonCharacter
ThirdPersonMap_Enemy_Feature

Usage
MakeNoise must be utilized to trigger employees to search for the original location of the noise.
isHidden can be found under the CanHideComponent and is currently solely set from the inspector.

Git-ClickUp
#865d5zjr3[Review]
#865d5zkej[Review]